--- contracts/UniswapV2Router03.sol
@@ -1,159 +1,0 @@
-pragma solidity =0.6.6;
-
-import './UniswapV2Router02.sol';
-import './libraries/SafeMath.sol';
-
-contract UniswapV2Router03 is UniswapV2Router02 {
-    using SafeMath for uint;
-
-    constructor(address _factory, address _WETH) public UniswapV2Router02(_factory, _WETH) {}
-
-    function removeLiquidityETH(
-        address token,
-        uint liquidity,
-        uint amountTokenMin,
-        uint amountETHMin,
-        address to,
-        uint deadline
-    ) public virtual override ensure(deadline) returns (uint amountToken, uint amountETH) {
-        (amountToken, amountETH) = removeLiquidity(
-            token,
-            WETH,
-            liquidity,
-            amountTokenMin,
-            amountETHMin,
-            address(this),
-            deadline
-        );
-        // use the current balance instead of amountToken to support fee-on-transfer tokens
-        TransferHelper.safeTransfer(token, to, IERC20(token).balanceOf(address(this)));
-        IWETH(WETH).withdraw(amountETH);
-        TransferHelper.safeTransferETH(to, amountETH);
-    }
-
-    // supports fee-on-transfer tokens
-    function _swapUsingObservedBalances(uint amountIn, address[] memory path, address _to)
-        internal
-        virtual
-        returns (uint[] memory amounts)
-    {
-        amounts = new uint[](path.length);
-        amounts[0] = amountIn;
-        for (uint i; i < path.length - 1; i++) {
-            (address input, address output) = (path[i], path[i + 1]);
-            IUniswapV2Pair pair = IUniswapV2Pair(UniswapV2Library.pairFor(factory, input, output));
-            uint amountInput;
-            uint amountOutput;
-            { // scope to avoid stack too deep errors
-            (uint reserveInput, uint reserveOutput) = UniswapV2Library.getReserves(factory, input, output);
-            // use the current balances - reserve to support fee-on-transfer tokens
-            amountInput = IERC20(input).balanceOf(address(pair)).sub(reserveInput);
-            amountOutput = UniswapV2Library.getAmountOut(amountInput, reserveInput, reserveOutput);
-            amounts[i + 1] = amountOutput;
-            }
-            (address token0,) = UniswapV2Library.sortTokens(input, output);
-            (uint amount0Out, uint amount1Out) = input == token0 ? (uint(0), amountOutput) : (amountOutput, uint(0));
-            address to = i < path.length - 2 ? UniswapV2Library.pairFor(factory, output, path[i + 2]) : _to;
-            pair.swap(amount0Out, amount1Out, to, new bytes(0));
-        }
-    }
-
-    function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
-        address[] calldata path,
-        address to,
-        uint deadline
-    ) external virtual override ensure(deadline) returns (uint[] memory amounts) {
-        TransferHelper.safeTransferFrom(
-            path[0], msg.sender, UniswapV2Library.pairFor(factory, path[0], path[1]), amountIn
-        );
-        amounts = _swapUsingObservedBalances(amountIn, path, to);
-        require(amounts[amounts.length - 1] >= amountOutMin, 'UniswapV2Router: INSUFFICIENT_OUTPUT_AMOUNT');
-    }
-    function swapTokensForExactTokens(
-        uint amountOut,
-        uint amountInMax,
-        address[] calldata path,
-        address to,
-        uint deadline
-    ) external virtual override ensure(deadline) returns (uint[] memory amounts) {
-        uint amountIn = UniswapV2Library.getAmountsIn(factory, amountOut, path)[0];
-        require(amountIn <= amountInMax, 'UniswapV2Router: EXCESSIVE_INPUT_AMOUNT');
-        TransferHelper.safeTransferFrom(
-            path[0], msg.sender, UniswapV2Library.pairFor(factory, path[0], path[1]), amountIn
-        );
-        amounts = _swapUsingObservedBalances(amountIn, path, to);
-        // note that if the path includes a fee-on-transfer token which takes a fee,
-        // there is _no_ guarantee that the received amount will equal amountOut
-    }
-    function swapExactETHForTokens(uint amountOutMin, address[] calldata path, address to, uint deadline)
-        external
-        override
-        payable
-        ensure(deadline)
-        returns (uint[] memory amounts)
-    {
-        require(path[0] == WETH, 'UniswapV2Router: INVALID_PATH');
-        uint amountIn = msg.value;
-        IWETH(WETH).deposit{value: amountIn}();
-        assert(IWETH(WETH).transfer(UniswapV2Library.pairFor(factory, path[0], path[1]), amountIn));
-        amounts = _swapUsingObservedBalances(amountIn, path, to);
-        require(amounts[amounts.length - 1] >= amountOutMin, 'UniswapV2Router: INSUFFICIENT_OUTPUT_AMOUNT');
-    }
-    function swapTokensForExactETH(uint amountOut, uint amountInMax, address[] calldata path, address to, uint deadline)
-        external
-        override
-        ensure(deadline)
-        returns (uint[] memory amounts)
-    {
-        require(path[path.length - 1] == WETH, 'UniswapV2Router: INVALID_PATH');
-        uint amountIn = UniswapV2Library.getAmountsIn(factory, amountOut, path)[0];
-        require(amountIn <= amountInMax, 'UniswapV2Router: EXCESSIVE_INPUT_AMOUNT');
-        TransferHelper.safeTransferFrom(
-            path[0], msg.sender, UniswapV2Library.pairFor(factory, path[0], path[1]), amountIn
-        );
-        amounts = _swapUsingObservedBalances(amountIn, path, address(this));
-        // use the current balance to support fee-on-transfer tokens
-        uint amountOutObserved = IERC20(WETH).balanceOf(address(this));
-        IWETH(WETH).withdraw(amountOutObserved);
-        TransferHelper.safeTransferETH(to, amountOutObserved);
-        // note that if the path includes a fee-on-transfer token which takes a fee,
-        // there is _no_ guarantee that the received amount will equal amountOut
-    }
-    function swapExactTokensForETH(uint amountIn, uint amountOutMin, address[] calldata path, address to, uint deadline)
-        external
-        override
-        ensure(deadline)
-        returns (uint[] memory amounts)
-    {
-        require(path[path.length - 1] == WETH, 'UniswapV2Router: INVALID_PATH');
-        TransferHelper.safeTransferFrom(
-            path[0], msg.sender, UniswapV2Library.pairFor(factory, path[0], path[1]), amountIn
-        );
-        amounts = _swapUsingObservedBalances(amountIn, path, address(this));
-        require(amounts[amounts.length - 1] >= amountOutMin, 'UniswapV2Router: INSUFFICIENT_OUTPUT_AMOUNT');
-        // use the current balance to support fee-on-transfer tokens
-        uint amountOutObserved = IERC20(WETH).balanceOf(address(this));
-        IWETH(WETH).withdraw(amountOutObserved);
-        TransferHelper.safeTransferETH(to, amountOutObserved);
-    }
-    function swapETHForExactTokens(uint amountOut, address[] calldata path, address to, uint deadline)
-        external
-        override
-        payable
-        ensure(deadline)
-        returns (uint[] memory amounts)
-    {
-        require(path[0] == WETH, 'UniswapV2Router: INVALID_PATH');
-        uint amountIn = UniswapV2Library.getAmountsIn(factory, amountOut, path)[0];
-        require(amountIn <= msg.value, 'UniswapV2Router: EXCESSIVE_INPUT_AMOUNT');
-        IWETH(WETH).deposit{value: amountIn}();
-        assert(IWETH(WETH).transfer(UniswapV2Library.pairFor(factory, path[0], path[1]), amountIn));
-        amounts = _swapUsingObservedBalances(amountIn, path, to);
-        // refund dust eth, if any
-        if (msg.value > amountIn) TransferHelper.safeTransferETH(msg.sender, msg.value - amountIn);
-        // note that if the path includes a fee-on-transfer token which takes a fee,
-        // there is _no_ guarantee that the received amount will equal amountOut
-    }
-}
