+++ test/UniswapV2Router02.spec.ts
@@ -1,10 +1,15 @@
 import chai, { expect } from 'chai'
-import { solidity, MockProvider, createFixtureLoader } from 'ethereum-waffle'
+import { solidity, MockProvider, createFixtureLoader, deployContract } from 'ethereum-waffle'
 import { Contract } from 'ethers'
-import { bigNumberify } from 'ethers/utils'
+import { BigNumber, bigNumberify } from 'ethers/utils'
 import { MaxUint256 } from 'ethers/constants'
+import IUniswapV2Pair from '@uniswap/v2-core/build/IUniswapV2Pair.json'
 
 import { v2Fixture } from './shared/fixtures'
+import { expandTo18Decimals, getApprovalDigest, MINIMUM_LIQUIDITY } from './shared/utilities'
+
+import DeflatingERC20 from '../build/DeflatingERC20.json'
+import { ecsign } from 'ethereumjs-util'
 
 chai.use(solidity)
 
@@ -12,118 +17,293 @@ const overrides = {
   gasLimit: 9999999
 }
 
-enum RouterVersion {
-  UniswapV2Router02 = 'UniswapV2Router02',
-  UniswapV2Router03 = 'UniswapV2Router03'
-}
+describe('UniswapV2Router02', () => {
+  const provider = new MockProvider({
+    hardfork: 'istanbul',
+    mnemonic: 'horn horn horn horn horn horn horn horn horn horn horn horn',
+    gasLimit: 9999999
+  })
+  const [wallet] = provider.getWallets()
+  const loadFixture = createFixtureLoader(provider, [wallet])
 
-describe('UniswapV2Router{02,03}', () => {
-  for (const routerVersion of Object.keys(RouterVersion)) {
-    const provider = new MockProvider({
-      hardfork: 'istanbul',
-      mnemonic: 'horn horn horn horn horn horn horn horn horn horn horn horn',
-      gasLimit: 9999999
-    })
-    const [wallet] = provider.getWallets()
-    const loadFixture = createFixtureLoader(provider, [wallet])
+  let token0: Contract
+  let token1: Contract
+  let router: Contract
+  beforeEach(async function() {
+    const fixture = await loadFixture(v2Fixture)
+    token0 = fixture.token0
+    token1 = fixture.token1
+    router = fixture.router02
+  })
 
-    let token0: Contract
-    let token1: Contract
-    let router: Contract
-    beforeEach(async function() {
-      const fixture = await loadFixture(v2Fixture)
-      token0 = fixture.token0
-      token1 = fixture.token1
-      router = {
-        [RouterVersion.UniswapV2Router02]: fixture.router02,
-        [RouterVersion.UniswapV2Router03]: fixture.router03
-      }[routerVersion as RouterVersion]
-    })
+  it('quote', async () => {
+    expect(await router.quote(bigNumberify(1), bigNumberify(100), bigNumberify(200))).to.eq(bigNumberify(2))
+    expect(await router.quote(bigNumberify(2), bigNumberify(200), bigNumberify(100))).to.eq(bigNumberify(1))
+    await expect(router.quote(bigNumberify(0), bigNumberify(100), bigNumberify(200))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_AMOUNT'
+    )
+    await expect(router.quote(bigNumberify(1), bigNumberify(0), bigNumberify(200))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
+    )
+    await expect(router.quote(bigNumberify(1), bigNumberify(100), bigNumberify(0))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
+    )
+  })
 
-    describe(routerVersion, () => {
-      it('quote', async () => {
-        expect(await router.quote(bigNumberify(1), bigNumberify(100), bigNumberify(200))).to.eq(bigNumberify(2))
-        expect(await router.quote(bigNumberify(2), bigNumberify(200), bigNumberify(100))).to.eq(bigNumberify(1))
-        await expect(router.quote(bigNumberify(0), bigNumberify(100), bigNumberify(200))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_AMOUNT'
-        )
-        await expect(router.quote(bigNumberify(1), bigNumberify(0), bigNumberify(200))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
-        )
-        await expect(router.quote(bigNumberify(1), bigNumberify(100), bigNumberify(0))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
-        )
-      })
+  it('getAmountOut', async () => {
+    expect(await router.getAmountOut(bigNumberify(2), bigNumberify(100), bigNumberify(100))).to.eq(bigNumberify(1))
+    await expect(router.getAmountOut(bigNumberify(0), bigNumberify(100), bigNumberify(100))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_INPUT_AMOUNT'
+    )
+    await expect(router.getAmountOut(bigNumberify(2), bigNumberify(0), bigNumberify(100))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
+    )
+    await expect(router.getAmountOut(bigNumberify(2), bigNumberify(100), bigNumberify(0))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
+    )
+  })
 
-      it('getAmountOut', async () => {
-        expect(await router.getAmountOut(bigNumberify(2), bigNumberify(100), bigNumberify(100))).to.eq(bigNumberify(1))
-        await expect(router.getAmountOut(bigNumberify(0), bigNumberify(100), bigNumberify(100))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_INPUT_AMOUNT'
-        )
-        await expect(router.getAmountOut(bigNumberify(2), bigNumberify(0), bigNumberify(100))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
-        )
-        await expect(router.getAmountOut(bigNumberify(2), bigNumberify(100), bigNumberify(0))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
-        )
-      })
+  it('getAmountIn', async () => {
+    expect(await router.getAmountIn(bigNumberify(1), bigNumberify(100), bigNumberify(100))).to.eq(bigNumberify(2))
+    await expect(router.getAmountIn(bigNumberify(0), bigNumberify(100), bigNumberify(100))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_OUTPUT_AMOUNT'
+    )
+    await expect(router.getAmountIn(bigNumberify(1), bigNumberify(0), bigNumberify(100))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
+    )
+    await expect(router.getAmountIn(bigNumberify(1), bigNumberify(100), bigNumberify(0))).to.be.revertedWith(
+      'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
+    )
+  })
 
-      it('getAmountIn', async () => {
-        expect(await router.getAmountIn(bigNumberify(1), bigNumberify(100), bigNumberify(100))).to.eq(bigNumberify(2))
-        await expect(router.getAmountIn(bigNumberify(0), bigNumberify(100), bigNumberify(100))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_OUTPUT_AMOUNT'
-        )
-        await expect(router.getAmountIn(bigNumberify(1), bigNumberify(0), bigNumberify(100))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
-        )
-        await expect(router.getAmountIn(bigNumberify(1), bigNumberify(100), bigNumberify(0))).to.be.revertedWith(
-          'UniswapV2Library: INSUFFICIENT_LIQUIDITY'
-        )
-      })
+  it('getAmountsOut', async () => {
+    await token0.approve(router.address, MaxUint256)
+    await token1.approve(router.address, MaxUint256)
+    await router.addLiquidity(
+      token0.address,
+      token1.address,
+      bigNumberify(10000),
+      bigNumberify(10000),
+      0,
+      0,
+      wallet.address,
+      MaxUint256,
+      overrides
+    )
 
-      it('getAmountsOut', async () => {
-        await token0.approve(router.address, MaxUint256)
-        await token1.approve(router.address, MaxUint256)
-        await router.addLiquidity(
-          token0.address,
-          token1.address,
-          bigNumberify(10000),
-          bigNumberify(10000),
-          0,
-          0,
-          wallet.address,
-          MaxUint256,
-          overrides
-        )
+    await expect(router.getAmountsOut(bigNumberify(2), [token0.address])).to.be.revertedWith(
+      'UniswapV2Library: INVALID_PATH'
+    )
+    const path = [token0.address, token1.address]
+    expect(await router.getAmountsOut(bigNumberify(2), path)).to.deep.eq([bigNumberify(2), bigNumberify(1)])
+  })
 
-        await expect(router.getAmountsOut(bigNumberify(2), [token0.address])).to.be.revertedWith(
-          'UniswapV2Library: INVALID_PATH'
-        )
-        const path = [token0.address, token1.address]
-        expect(await router.getAmountsOut(bigNumberify(2), path)).to.deep.eq([bigNumberify(2), bigNumberify(1)])
-      })
+  it('getAmountsIn', async () => {
+    await token0.approve(router.address, MaxUint256)
+    await token1.approve(router.address, MaxUint256)
+    await router.addLiquidity(
+      token0.address,
+      token1.address,
+      bigNumberify(10000),
+      bigNumberify(10000),
+      0,
+      0,
+      wallet.address,
+      MaxUint256,
+      overrides
+    )
 
-      it('getAmountsIn', async () => {
-        await token0.approve(router.address, MaxUint256)
-        await token1.approve(router.address, MaxUint256)
-        await router.addLiquidity(
-          token0.address,
-          token1.address,
-          bigNumberify(10000),
-          bigNumberify(10000),
-          0,
-          0,
-          wallet.address,
-          MaxUint256,
-          overrides
-        )
+    await expect(router.getAmountsIn(bigNumberify(1), [token0.address])).to.be.revertedWith(
+      'UniswapV2Library: INVALID_PATH'
+    )
+    const path = [token0.address, token1.address]
+    expect(await router.getAmountsIn(bigNumberify(1), path)).to.deep.eq([bigNumberify(2), bigNumberify(1)])
+  })
+})
 
-        await expect(router.getAmountsIn(bigNumberify(1), [token0.address])).to.be.revertedWith(
-          'UniswapV2Library: INVALID_PATH'
-        )
-        const path = [token0.address, token1.address]
-        expect(await router.getAmountsIn(bigNumberify(1), path)).to.deep.eq([bigNumberify(2), bigNumberify(1)])
-      })
+describe('fee-on-transfer tokens', () => {
+  const provider = new MockProvider({
+    hardfork: 'istanbul',
+    mnemonic: 'horn horn horn horn horn horn horn horn horn horn horn horn',
+    gasLimit: 9999999
+  })
+  const [wallet] = provider.getWallets()
+  const loadFixture = createFixtureLoader(provider, [wallet])
+
+  let DTT: Contract
+  let WETH: Contract
+  let router: Contract
+  let pair: Contract
+  beforeEach(async function() {
+    const fixture = await loadFixture(v2Fixture)
+
+    WETH = fixture.WETH
+    router = fixture.router02
+
+    DTT = await deployContract(wallet, DeflatingERC20, [expandTo18Decimals(10000)])
+
+    // make a DTT<>WETH pair
+    await fixture.factoryV2.createPair(DTT.address, WETH.address)
+    const pairAddress = await fixture.factoryV2.getPair(DTT.address, WETH.address)
+    pair = new Contract(pairAddress, JSON.stringify(IUniswapV2Pair.abi), provider).connect(wallet)
+  })
+
+  afterEach(async function() {
+    expect(await provider.getBalance(router.address)).to.eq(0)
+  })
+
+  async function addLiquidity(DTTAmount: BigNumber, WETHAmount: BigNumber) {
+    await DTT.approve(router.address, MaxUint256)
+    await router.addLiquidityETH(DTT.address, DTTAmount, DTTAmount, WETHAmount, wallet.address, MaxUint256, {
+      ...overrides,
+      value: WETHAmount
     })
   }
+
+  it('removeLiquidityETHSupportingFeeOnTransferTokens', async () => {
+    const DTTAmount = expandTo18Decimals(1)
+    const ETHAmount = expandTo18Decimals(4)
+    await addLiquidity(DTTAmount, ETHAmount)
+
+    const DTTInPair = await DTT.balanceOf(pair.address)
+    const WETHInPair = await WETH.balanceOf(pair.address)
+    const liquidity = await pair.balanceOf(wallet.address)
+    const totalSupply = await pair.totalSupply()
+    const NaiveDTTExpected = DTTInPair.mul(liquidity).div(totalSupply)
+    const WETHExpected = WETHInPair.mul(liquidity).div(totalSupply)
+
+    await pair.approve(router.address, MaxUint256)
+    await router.removeLiquidityETHSupportingFeeOnTransferTokens(
+      DTT.address,
+      liquidity,
+      NaiveDTTExpected,
+      WETHExpected,
+      wallet.address,
+      MaxUint256,
+      overrides
+    )
+  })
+
+  it('removeLiquidityETHWithPermitSupportingFeeOnTransferTokens', async () => {
+    const DTTAmount = expandTo18Decimals(1)
+      .mul(100)
+      .div(99)
+    const ETHAmount = expandTo18Decimals(4)
+    await addLiquidity(DTTAmount, ETHAmount)
+
+    const expectedLiquidity = expandTo18Decimals(2)
+
+    const nonce = await pair.nonces(wallet.address)
+    const digest = await getApprovalDigest(
+      pair,
+      { owner: wallet.address, spender: router.address, value: expectedLiquidity.sub(MINIMUM_LIQUIDITY) },
+      nonce,
+      MaxUint256
+    )
+    const { v, r, s } = ecsign(Buffer.from(digest.slice(2), 'hex'), Buffer.from(wallet.privateKey.slice(2), 'hex'))
+
+    const DTTInPair = await DTT.balanceOf(pair.address)
+    const WETHInPair = await WETH.balanceOf(pair.address)
+    const liquidity = await pair.balanceOf(wallet.address)
+    const totalSupply = await pair.totalSupply()
+    const NaiveDTTExpected = DTTInPair.mul(liquidity).div(totalSupply)
+    const WETHExpected = WETHInPair.mul(liquidity).div(totalSupply)
+
+    await pair.approve(router.address, MaxUint256)
+    await router.removeLiquidityETHWithPermitSupportingFeeOnTransferTokens(
+      DTT.address,
+      liquidity,
+      NaiveDTTExpected,
+      WETHExpected,
+      wallet.address,
+      MaxUint256,
+      false,
+      v,
+      r,
+      s,
+      overrides
+    )
+  })
+
+  describe('swapExactTokensForTokensSupportingFeeOnTransferTokens', () => {
+    const DTTAmount = expandTo18Decimals(5)
+      .mul(100)
+      .div(99)
+    const ETHAmount = expandTo18Decimals(10)
+    const amountIn = expandTo18Decimals(1)
+
+    beforeEach(async () => {
+      await addLiquidity(DTTAmount, ETHAmount)
+    })
+
+    it('DTT -> WETH', async () => {
+      await DTT.approve(router.address, MaxUint256)
+
+      await router.swapExactTokensForTokensSupportingFeeOnTransferTokens(
+        amountIn,
+        0,
+        [DTT.address, WETH.address],
+        wallet.address,
+        MaxUint256,
+        overrides
+      )
+    })
+
+    // WETH -> DTT
+    it('WETH -> DTT', async () => {
+      await WETH.deposit({ value: amountIn }) // mint WETH
+      await WETH.approve(router.address, MaxUint256)
+
+      await router.swapExactTokensForTokensSupportingFeeOnTransferTokens(
+        amountIn,
+        0,
+        [WETH.address, DTT.address],
+        wallet.address,
+        MaxUint256,
+        overrides
+      )
+    })
+  })
+
+  // ETH -> DTT
+  it('swapExactETHForTokensSupportingFeeOnTransferTokens', async () => {
+    const DTTAmount = expandTo18Decimals(10)
+      .mul(100)
+      .div(99)
+    const ETHAmount = expandTo18Decimals(5)
+    const swapAmount = expandTo18Decimals(1)
+    await addLiquidity(DTTAmount, ETHAmount)
+
+    await router.swapExactETHForTokensSupportingFeeOnTransferTokens(
+      0,
+      [WETH.address, DTT.address],
+      wallet.address,
+      MaxUint256,
+      {
+        ...overrides,
+        value: swapAmount
+      }
+    )
+  })
+
+  // DTT -> ETH
+  it('swapExactTokensForETHSupportingFeeOnTransferTokens', async () => {
+    const DTTAmount = expandTo18Decimals(5)
+      .mul(100)
+      .div(99)
+    const ETHAmount = expandTo18Decimals(10)
+    const swapAmount = expandTo18Decimals(1)
+
+    await addLiquidity(DTTAmount, ETHAmount)
+    await DTT.approve(router.address, MaxUint256)
+
+    await router.swapExactTokensForETHSupportingFeeOnTransferTokens(
+      swapAmount,
+      0,
+      [DTT.address, WETH.address],
+      wallet.address,
+      MaxUint256,
+      overrides
+    )
+  })
 })
