+++ test/shared/utilities.ts
@@ -2,6 +2,8 @@ import { Contract } from 'ethers'
 import { Web3Provider } from 'ethers/providers'
 import { BigNumber, bigNumberify, keccak256, defaultAbiCoder, toUtf8Bytes, solidityPack } from 'ethers/utils'
 
+export const MINIMUM_LIQUIDITY = bigNumberify(10).pow(3)
+
 const PERMIT_TYPEHASH = keccak256(
   toUtf8Bytes('Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)')
 )
